requirements.txt
torch~=2.4.1
torchvision~=0.19.1
pandas~=2.2.3
Pillow~=10.4.0
tqdm~=4.66.5